scan and re direct to upi apps option enabled

+++ ExpenseTracker/app/src/main/java/com/expensetracker/app/data/model/CategorySpend.kt
@@ -9,9 +9,13 @@ data class CategorySpend(
     val spentMinor: Long,
     val limitMinor: Long?,
 ) {
-    /** 0f..1f+ ; not coerced so callers can distinguish "at limit" from "over limit". */
+    /** 0f..1f+ ; not coerced so callers can distinguish "at limit" from "over limit".
+     * Divides as Double, not Float — see LimitAlertEvaluator's tierFor for why: Float only
+     * represents integers exactly up to ~16.7M, so a Float division here could round onto the
+     * wrong side of a color threshold for spend/limits beyond that while the (Double-based)
+     * notification logic reads the same numbers correctly. */
     val progress: Float
-        get() = if (limitMinor != null && limitMinor > 0) spentMinor.toFloat() / limitMinor.toFloat() else 0f
+        get() = if (limitMinor != null && limitMinor > 0) (spentMinor.toDouble() / limitMinor.toDouble()).toFloat() else 0f
 }
 
 @Immutable
@@ -28,5 +32,5 @@ data class MonthlySummary(
     val categorySpends: List<CategorySpend>,
 ) {
     val overallProgress: Float
-        get() = if (overallLimitMinor != null && overallLimitMinor > 0) totalSpentMinor.toFloat() / overallLimitMinor.toFloat() else 0f
+        get() = if (overallLimitMinor != null && overallLimitMinor > 0) (totalSpentMinor.toDouble() / overallLimitMinor.toDouble()).toFloat() else 0f
 }
